Fix ups post rename
